update db + crud nhomsanpham, sanpham

+++ database/migrations/2021_11_01_130708_sanpham.php
@@ -18,8 +18,8 @@ class Sanpham extends Migration
             $table->bigInteger('nhomsanphamid')->unsigned();
             $table->string('ten', 100);
             $table->text('mota')->nullable();
-            $table->float('gia');
-            $table->float('giaban')->nullable();
+            $table->decimal('gia',15,3);
+            $table->decimal('giaban',15,3)->nullable();
             $table->string('anh', 255)->nullable();
             $table->string('danhsachanh', 255)->nullable();
             $table->tinyInteger('trangthai')->default(1)->comment('1-public, 0-private');
